requirements.txt
django==2.2
django-pure-pagination
mysqlclient
pillow